account-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
kafka-python
cryptography
httpx==0.28.1  